add APIConvertible protocol
